// src/app/models/taller.ts
export class Taller {
    id?: number;
    oficina?: string;
    ciudad?: string;
    ciudadID?: number;
    estado?: string;
    nombre?: string;
    nit?: string;
    direccion?: string;
    marcaTallerID?: number;
    clasesVehiculo?: string;
    telefono?: string;
    horariosAtencion?: string;

    // TODO: crear nuevo modelo
    marcaID?: number;
    segmento?: string;
    placa?: string;
    latitud?: string;
    longitud?: string;
}